Remove pre-created webapps/static directory.
The created openhab-runtime debian package would like to create
a symlink /usr/share/openhab/webapps/static that points to
/var/lib/openhab/webapps/static.
If the directory already exists because it is part of the data packed,
this will result in a missing symlink.

The directory /usr/share/openhab/webapps/static is owned by root and
must not be changed on runtime (so not writable by openhab user/group).
This is already done, but result in errors at runtime.

So we have to guarantee that webapps/static is a symlink to
/var/lib/openhab/webapps/static and so writable on runtime by openhab
user.