Z-Wave: Node Deletion & Add Controller IsFailedNode
A Z-Wave node can only be deleted if its present in the controllers
"Failed List" resident in its own NVRAM.
Added the Z-Wave IsFailedNode Message Class.
Added new Failed status if the controller has it flagged as failed in
its own database.
This then allows a node to be deleted.
Visibility of the "Delete Node" button uses this new logic.