bump java environment to fix annotation compliance
Well, there is a @Override annotation on one of the methods, but it
doesn’t have a superclass–it is annotating a method that is being
implemented (from an interface), not a method that is inherited (from a
superclass).

The @Override annotation is a Java 1.5 feature, and the error message I
was getting was coming (correctly) from a 1.5 compiler. However in Java
1.6 the @Override annotation can also be applied to methods that are
implemented from an interface.

Quote:
http://www.filsa.net/2008/02/11/maven-and-the-method-does-not-override-a-method-from-its-superclass-error/