add missing imports (for bundled jar files)
The OSGi specification mandates that a bundle declare all package
depenencies using either Import-Package or Require-Bundle. The one
exception to this rule is the java.* packages which is always delegated
to the boot classpath. All other packages dependencies must be declared
in the bundle's manifest file.

This OSGi bundle bundles a xstream jar file that is added to the
classpath. So the MANIFEST.MF of the xstream jar is not evaluated by the
OSGi framework. We have to add the packages, that are necessary for the used
xstream functionality to the import section of the MANIFEST.MF of this bundle.

Signed-off-by: Markus Rathgeb <maggu2810@gmail.com>