MapDB Persistence: PercentType restore needs to allow Decimal values.
In openHAB, the PercentType value is derived from DecimalType.  The
MapDB Persistence is pre-parsing the value, assuming it's an Integer,
and failing when it wasn't an Integer.

In some cases, this would cause the startup process to enter a never
ending loop, since the Persistence-Restore functionality was failing.

With this change, we simple hand the String value to
PercentType.valueOf() and let it perform any necessary conversions.

Signed-off-by: Mark Clark <mr.guessed@gmail.com>