MiOS Binding - Implement OH2 feature.xml for mios*.map transform files.
The MiOS Binding uses a number of openHAB MAP Transformation files to
perform it's work.  In openHAB 1.x, these were manually copied by users
into the `configurations/transform` directory on their openHAB runtime.

openHAB 2.x uses Karaf to aid in the Binding deployment steps, and uses
`feature.xml` as a configuration file:

https://karaf.apache.org/manual/latest/users-guide/provisioning.html

This change implements the extra `feature.xml` entries for the MiOS
Binding, so the MAP Transformation files can be picked up when the
MiOS Binding is deployed to openHAB 2.x

Signed-off-by: Mark Clark <mr.guessed@gmail.com> (github: mrguessed)