ONKYO BINDING MULTIPLE DEVICE STATUS UPDATE FIX
Fixes #4255 regarding problems with status updates when multiple devices are configured. There was no check to match the device ID against the device ID in the binding specification, so status updates from devices were being posted to items bound to different devices. A full comparison (instead of the first three characters) was also added for advanced command keys as three characters is not enough to distinguish advanced commands. This caused status updates to be written to the wrong item.