[mios] Improve logging output of MiOS Item Generator output.
The MiOS Item Generator builds an openHAB Items file using the metadata
from a MiOS/Vera Unit.

In some cases, users were confused by the output of the tool and this
lead to issues in the subsequent configuration steps.

This PR adds further logging/output to guide the user in the
configuration.

Signed-off-by: Mark Clark <mr.guessed@gmail.com> (github: mrguessed)