Sonance binding: volume direct IP commands (#4429)
* Sonance binding: changed volume commands to direct IP commands

The manifacture recently added direct IP absolute volume commands, so
those are implemented. It required a specific firmware as described in
the Wiki of the binding.
Signed-off-by: Laurens Van Acker laurens-at-van-dot-acker-dot-be

* Sonance binding: changed volume commands to direct IP commands

The manifacture recently added direct IP absolute volume commands, so
those are implemented. It required a specific firmware as described in
the Wiki of the binding.
Signed-off-by: Laurens Van Acker laurens-at-van-dot-acker-dot-be

* Changes based on comments from Watou

* Correct copyright header en fixed {@inheritDoc} tags

* Added functions for OH2 and a new line at the end