Add option to send plain literals messages.
The TCP binding does not allow to send plain literal strings straight from the item declaration <transformationrule> (the 4th argument in brackets), but forces to go through transformations.
A simple way to achieve that without changing the code much is to treat an 'empty' transformation to pass the argument directly as the response: instead of "MAP(my.device.map)" or "REGEX((.*))" for example, just type "(1234ABC)" to send the string 1234ABC. The parenthesis are needed to match the pattern tested before, otherwise much more code needs to changed. Also we can assume that no transformation rule will ever be named with an empty string, but it still matches the pattern.