[modbus] Harmonize transaction id handling in serial to match TCP
Now when requests are retried, the transaction id is incremented every
time. I think it’s always safe to increase the id, but leaving it as
is might be risky (some slaves might not like it, for example).

Increasing the ID is more like “fire and forget” kind of behaviour, if
the original request fails, we make completely new one instead of
repeating the old one.