Update GPIOGenericBindingProvider.java (#5176)
* Update GPIOGenericBindingProvider.java

I'm proposing additional property in the GPIO binding item configuration which defines the initial value of the GPIO pin. With the old implementation it was always defined initially to LOW (0). There are cases when it is necessary the initial value to be HIGH (1). There is support for this in org.openhab.io.gpio but wasn't implemented in the configuration part of the GPIO binding.

* Update README.md

change requested from review