[fritzboxtr064] Fetch multiple values in one fritzbox request (#5221)
Rework of the ItemMap concept, which maps a configured item to a call of
a FritzBox SOAP service. The previous ItemMap only supported a single
response value per SOAP call. The new MultiItemMap can extract values
for multiple configured items from a single SOAP service response. This
avoids multiple calls to the same service if several items are
configured.

This change also improves and unifies the  handling of SOAP faults and
missing values.

Signed-off-by: Michael Koch <tensberg@gmx.net> (github: tensberg)