KNX1_commandGAbyReceiveCommand2KNX-stateGAbyReceiveUpdate2KNX
Now commandGAs are sent only by internalReceiveCommand to KNX
and stateGAs are sent only by internalReceiveUpdate to KNX

It is fully backward compatible to older KNX1 binding versions, in OH1
it works without changes.

In OH2 Rules, now only sendCommand has to be used, to send a value
directly to KNX bus. Using postUpdate in Rules does not send the value
to KNX bus anymore.

To continue using the old faulty mode in OH2, the configuration
parameter "environmentModeFelix = false" can be set.