[FritzboxTr064] Unify read/write input argument handling
All items use the same item argument name (=XML element name) for
reading and writing the value of the item. They also use the same
additional input arguments for both reading and writing. This
refactoring replaces the separate initialization of read and
write parameters with a shared item argument name and config input
arguments list.

"parameter" is renamed to "argument", because this is the word used in
the FritzBox TR064 documentation.

The item argument name is given in the ItemMap constructor. For read
calls, the value is extracted from the XML response and parsed by the
SoapValueParser. For write calls, the argument is part of the request,
its value is derived from the OpenHab Command.

Signed-off-by: Michael Koch <tensberg@gmx.net> (github: tensberg)