Filters now allow any character, but require single quotes
This change makes single quotes around the filters a requirement, which is what the documentation already shows. Single quotes inside the filter will stop the matching, so they should be replaced with a period (.) in the Item definition. The benefit of this change is that the characters in the filters no longer have any other limitation. Previously, they were limited, [as discussed in the forum](https://community.openhab.org/t/caldav-personal-regular-expression/84751/13?u=5iver).

Signed-off-by: Scott Rushworth <openhab@5iver.com>