Replaced "must not" by "need not"
I guess it had been a translation error from german - saying "must not be ordered" means it has to be be *not* in order. That doesn't make sense to me. Guessing from the german wording "muss nicht sortiert sein" I guess the meaning should be that it may be or may be not sorted - which is "need not be ordered" in english.